+++ Neo4j.Driver/Neo4j.Driver.Tests/Internal/BookmarkManager/BookmarkManagerFactoryTests.cs
@@ -20,6 +20,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit;
 using FluentAssertions;
+using Neo4j.Driver.Experimental;
 
 namespace Neo4j.Driver.Internal.BookmarkManager;
