+++ Neo4j.Driver/Neo4j.Driver.Tests/Internal/BookmarkManager/DefaultBookmarkManagerTests.cs
@@ -21,6 +21,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Moq;
+using Neo4j.Driver.Experimental;
 using Xunit;
 
 namespace Neo4j.Driver.Internal.BookmarkManager;
