+++ Neo4j.Driver/Neo4j.Driver/Experimental/IBookmarkManagerFactory.cs
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Neo4j.Driver;
+namespace Neo4j.Driver.Experimental;
 
 /// <summary>
 /// Factory for <see cref="IBookmarkManager"/>.
