+++ Neo4j.Driver/Neo4j.Driver/Internal/DefaultBookmarkManager.cs
@@ -20,6 +20,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Neo4j.Driver.Experimental;
 
 namespace Neo4j.Driver.Internal;
