+++ Neo4j.Driver/Neo4j.Driver/SessionConfig.cs
@@ -18,6 +18,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Neo4j.Driver.Experimental;
 using Neo4j.Driver.Internal;
 
 namespace Neo4j.Driver
@@ -126,7 +127,7 @@ namespace Neo4j.Driver
             internal set => _impersonatedUser = (!string.IsNullOrEmpty(value)) ? value : throw new ArgumentNullException();
         }
 
-        public IBookmarkManager BookmarkManager { get; set; }
+        internal IBookmarkManager BookmarkManager { get; set; }
     }
 
     /// <summary>
